+++ src/main/java/com/accsaber/backend/model/dto/APResult.java
@@ -1,6 +1,4 @@
 package com.accsaber.backend.model.dto;
 
-import java.math.BigDecimal;
-
-public record APResult(BigDecimal rawAP, BigDecimal normalizedAP) {
+public record APResult(double rawAP, double normalizedAP) {
 }
