+++ src/main/java/com/accsaber/backend/model/dto/request/curve/CreateCurveRequest.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.request.curve;
 
-import java.math.BigDecimal;
 
 import com.accsaber.backend.model.entity.CurveType;
 
@@ -20,14 +19,14 @@ public class CreateCurveRequest {
     private String formula;
 
     private String xParameterName;
-    private BigDecimal xParameterValue;
+    private Double xParameterValue;
 
     private String yParameterName;
-    private BigDecimal yParameterValue;
+    private Double yParameterValue;
 
     private String zParameterName;
-    private BigDecimal zParameterValue;
+    private Double zParameterValue;
 
-    private BigDecimal scale;
-    private BigDecimal shift;
+    private Double scale;
+    private Double shift;
 }
