+++ src/main/java/com/accsaber/backend/model/dto/request/item/UpdateItemModifierRequest.java
@@ -1,13 +1,12 @@
 package com.accsaber.backend.model.dto.request.item;
 
-import java.math.BigDecimal;
 
 import lombok.Data;
 
 @Data
 public class UpdateItemModifierRequest {
 
-    private BigDecimal globalDropChance;
+    private Double globalDropChance;
     private String seasonStart;
     private String seasonEnd;
 }
