+++ src/main/java/com/accsaber/backend/model/dto/response/ModifierResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.Builder;
@@ -13,5 +12,5 @@ public class ModifierResponse {
     UUID id;
     String name;
     String code;
-    BigDecimal multiplier;
+    Double multiplier;
 }
